Corecciones de archivos y pruebas unitarias
