+++ BAZProject/BAZProjectTests/TredingTests/TrendingViewModelTests.swift
@@ -51,7 +51,7 @@ final class TrendingViewModelTests: XCTestCase {
         //When
         movieAPIMock?.movies = movies
         sut?.bindMovies { expectation.fulfill() }
-        sut?.getmovies(category: .NowPlaying)
+        sut?.getmovies(category: .nowPlaying)
         wait(for: [expectation], timeout: 0.1)
         let count = sut?.getMovieCount()
         //Then
