+++ BAZProject/Commons/Commons.swift
@@ -57,30 +57,3 @@ enum StringsTitles: String {
     case search = "buscar"
     case treding = "Trending"
 }
-//MARK: - Extensiones
-
-// Metodos extras para el uso de peliculas con Strings
-extension String {
-    /// Formato del nombre de la pelicula para poder buscarla
-    func formatterMovieName() -> String {
-        self.replacingOccurrences(of: " ", with: "%20", options: NSString.CompareOptions.literal, range: nil) //Dividir palabras para la busqueda
-    }
-}
-
-// Metodos extras para la colocacion de las imagenes
-extension UIImageView {
-    /// Consulta desde una url para colocar la imagen
-    func loadImage(url: URL) -> URLSessionDownloadTask {
-        let session = URLSession.shared
-        let downloadTask = session.downloadTask(with: url){[weak self] url, response, error in
-            guard let self = self else { return }
-            if error == nil, let url = url, let data = try? Data(contentsOf: url), let image = UIImage(data: data){
-                DispatchQueue.main.async {
-                    self.image = image
-                }
-            }
-        }
-        downloadTask.resume()
-        return downloadTask
-    }
-}
